+++ shopos-theme/functions.php
@@ -7,7 +7,7 @@
 
 defined( 'ABSPATH' ) || exit;
 
-define( 'SHOPOS_THEME_VERSION',   '1.13.1' );
+define( 'SHOPOS_THEME_VERSION',   '1.14.0' );
 define( 'SHOPOS_THEME_PATH',      get_stylesheet_directory() );
 define( 'SHOPOS_THEME_URL',       get_stylesheet_directory_uri() );
 define( 'SHOPOS_THEME_ASSETS',    SHOPOS_THEME_URL . '/assets' );
@@ -19,7 +19,9 @@ require_once SHOPOS_THEME_PATH . '/inc/hooks.php';
 require_once SHOPOS_THEME_PATH . '/inc/woocommerce.php';
 require_once SHOPOS_THEME_PATH . '/inc/customizer.php';
 require_once SHOPOS_THEME_PATH . '/inc/design-tokens.php';
+require_once SHOPOS_THEME_PATH . '/inc/class-shopos-template-loader.php';
 require_once SHOPOS_THEME_PATH . '/inc/updater.php';
 
 ShopOS_Theme::instance();
+ShopOS_Theme_Template_Loader::instance()->register();
 ShopOS_Theme_Plugin_Dependencies::instance();